[frontend] 1205 - Delete an item from a list (inject, atomic testing, scenario, simlation) Improved Atomictesting page loading
