+++ openbas-front/src/admin/components/common/injects/UpdateInject.tsx
@@ -1,35 +1,31 @@
 import React, { useEffect, useRef, useState } from 'react';
 import Drawer from '../../../../components/common/Drawer';
 import { useFormatter } from '../../../../components/i18n';
-import type { Inject } from '../../../../utils/api-types';
-import useDataLoader from '../../../../utils/hooks/useDataLoader';
-import { useAppDispatch } from '../../../../utils/hooks';
+import type { Inject, InjectResultDTO } from '../../../../utils/api-types';
 import UpdateInjectDetails from './UpdateInjectDetails';
 import type { TeamStore } from '../../../../actions/teams/Team';
-import { fetchInject } from '../../../../actions/Inject';
-import { useHelper } from '../../../../store';
-import type { InjectHelper } from '../../../../actions/injects/inject-helper';
 
 interface Props {
   open: boolean;
   handleClose: () => void;
   onUpdateInject: (data: Inject) => Promise<void>;
+  inject: InjectResultDTO;
   injectId: string;
   isAtomic?: boolean;
   teamsFromExerciseOrScenario: TeamStore[];
 }
 
-const UpdateInject: React.FC<Props> = ({ open, handleClose, onUpdateInject, injectId, isAtomic = false, teamsFromExerciseOrScenario, ...props }) => {
+const UpdateInject: React.FC<Props> = ({
+  open,
+  handleClose,
+  onUpdateInject,
+  inject,
+  injectId,
+  isAtomic = false,
+  teamsFromExerciseOrScenario,
+  ...props }) => {
   const { t } = useFormatter();
-  const dispatch = useAppDispatch();
   const drawerRef = useRef(null);
-  // Fetching data
-  const { inject } = useHelper((helper: InjectHelper) => ({
-    inject: helper.getInject(injectId),
-  }));
-  useDataLoader(() => {
-    dispatch(fetchInject(injectId));
-  });
 
   const [injectorContract, setInjectorContract] = useState(null);
   useEffect(() => {
